requirements.txt
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.0.8
Flask-Bcrypt==0.7.1
flask-blacklist==0.0.1
Flask-Cors==3.0.8
Flask-DebugToolbar==0.11.0
Flask-JWT-Extended==3.25.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-Moment==0.10.0
flask-mongoengine==1.0.0
Flask-MySQLdb==0.2.0
Flask-OAuthlib==0.9.5
Flask-RESTful==0.3.7
flask-sieve==2.0.1
Flask-SQLAlchemy==2.4.4
Flask-SSLify==0.1.5
Flask-WTF==0.14.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
mongoengine==0.22.1
mysqlclient==2.0.3
oauthlib==2.1.0
psycopg2==2.8.4
pymongo==3.11.0
python-dateutil==2.8.1
requests==2.22.0
resampy==0.2.2
rsa==4.6
SQLAlchemy==1.3.23
SQLAlchemy-Searchable==1.1.0
SQLAlchemy-Utils==0.36.7
typing-extensions==3.10.0.1
validators==0.15.0
w3lib==1.22.0
websockets==8.1
Werkzeug==0.16.0
uwsgi
gunicorn==20.0.4
python-dotenv==0.10.3
path==16.3.0